Use long doubles

+++ thermodynamic_process.cpp
@@ -1,3 +1,3 @@
 #include "thermodynamic_process.h"
 
-ThermodynamicProcess::ThermodynamicProcess(double const nu) : nu_(nu) {}
+ThermodynamicProcess::ThermodynamicProcess(long double const nu) : nu_(nu) {}
